move app settings types to types package
