+++ packages/renderer/src/renderer/contexts/settings/SettingFlags/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { SettingItem } from '@app/screens/Settings/types';
+import type { SettingItem } from '@polkadot-live/types/settings';
 
 export interface SettingFlagsContextInterface {
   getSwitchState: (setting: SettingItem) => boolean;
